Migrate docs demos to OpenUI Cloud
