+++ docs/.env.example
@@ -1,0 +1,12 @@
+# Hosted workloads are available as soon as their dedicated key is non-empty.
+# Use a separately named OpenUI Cloud key for every workload so usage is attributable by key.
+THESYS_API_KEY_DOCS_CHAT_OSS=
+THESYS_API_KEY_DOCS_CHAT_CLOUD=
+THESYS_API_KEY_DOCS_COMPARE_MARKDOWN=
+THESYS_API_KEY_DOCS_COMPARE_OSS=
+THESYS_API_KEY_DOCS_COMPARE_CLOUD=
+THESYS_API_KEY_DOCS_PLAYGROUND=
+THESYS_API_KEY_DOCS_GITHUB=
+
+# Optional origin allowlist used by the managed Cloud surfaces.
+OPENUI_CLOUD_ALLOWED_ORIGINS=
