+++ docs/app/api/compare/markdown/route.ts
@@ -1,0 +1,6 @@
+import type { NextRequest } from "next/server";
+import { handleChatCompletions } from "../../_lib/chat-completions-handler";
+
+export function POST(request: NextRequest) {
+  return handleChatCompletions(request, "compare-markdown", "markdown");
+}
