+++ docs/app/api/demo/github/stream/route.ts
@@ -2,7 +2,8 @@ import {
   createDemoCreditsExhaustedResponse,
   isDemoCreditsExhaustedError,
 } from "@/lib/demo-credits";
-import { BASE_URL } from "@/lib/source";
+import { readOpenuiCloudConfig } from "@/lib/openui-cloud/config";
+import { unavailableResponse } from "@/lib/openui-cloud/errors";
 import { generatePrompt, type PromptSpec } from "@openuidev/lang-core";
 import { readFileSync } from "fs";
 import { type NextRequest } from "next/server";
@@ -60,21 +61,14 @@ export async function POST(req: NextRequest) {
   }
   chatMessages.push({ role: "user", content: prompt });
 
-  const apiKey = process.env.OPENROUTER_API_KEY;
-  if (!apiKey) {
-    return Response.json(
-      { error: { message: "OPENROUTER_API_KEY not configured" } },
-      { status: 500 },
-    );
-  }
+  const config = readOpenuiCloudConfig("github");
+  if (!config) return unavailableResponse();
 
-  const res = await fetch("https://openrouter.ai/api/v1/chat/completions", {
+  const res = await fetch(`${config.embedBaseUrl}/chat/completions`, {
     method: "POST",
     headers: {
-      Authorization: `Bearer ${apiKey}`,
+      Authorization: `Bearer ${config.apiKey}`,
       "Content-Type": "application/json",
-      "HTTP-Referer": `${BASE_URL}/demo/github`,
-      "X-Title": "OpenUI GitHub Demo",
     },
     body: JSON.stringify({
       model: GITHUB_DEMO_MODEL,
@@ -93,7 +87,7 @@ export async function POST(req: NextRequest) {
     return Response.json(
       {
         error: (err as { error?: { message?: string } }).error ?? {
-          message: `OpenRouter error ${res.status}`,
+          message: `OpenUI Cloud error ${res.status}`,
         },
       },
       { status: res.status },
@@ -102,7 +96,7 @@ export async function POST(req: NextRequest) {
 
   // Keep credit handling to provider 4xx responses. Provider-specific mid-stream
   // error chunks are intentionally passed through because they are harder to
-  // maintain across OpenRouter/OpenAI streaming shape changes.
+  // maintain across OpenUI Cloud/OpenAI streaming shape changes.
   return new Response(res.body, {
     headers: {
       "Content-Type": "text/event-stream",
