+++ docs/app/api/openui-cloud/compare/chat/route.ts
@@ -1,0 +1,5 @@
+import { handleCloudResponses } from "../../_lib/responses-handler";
+
+export function POST(request: Request) {
+  return handleCloudResponses(request, "compare-cloud");
+}
