+++ docs/app/api/openui-cloud/compare/frontend-token/route.ts
@@ -1,0 +1,5 @@
+import { handleFrontendToken } from "../../_lib/frontend-token-handler";
+
+export function POST(request: Request) {
+  return handleFrontendToken(request, "compare-cloud");
+}
