+++ docs/app/demos/constants.ts
@@ -3,10 +3,14 @@ export type Status = "idle" | "streaming" | "done" | "error";
 
 export const MODELS = [
   "anthropic/claude-sonnet-4.6",
-  "anthropic/claude-haiku-4.5",
+  "google/gemini-3.1-pro-free",
   "openai/gpt-5.2",
 ] as const;
 export type Model = (typeof MODELS)[number];
+
+export function isPlaygroundModel(value: unknown): value is Model {
+  return typeof value === "string" && (MODELS as readonly string[]).includes(value);
+}
 
 export const STARTER_PROMPTS = [
   "Weather dashboard",
